added main-interface draft for pocket, fixed transactions package
