+++ packages/main-interface/.ctirc
@@ -1,0 +1,31 @@
+// created from 'create-ts-index'
+
+{
+  addNewline: true,
+  excludes: [
+    '@types',
+    'typings',
+    '__test__',
+    '__tests__',
+    'dist',
+    'node_modules',
+    'proto'
+  ],
+  fileExcludePatterns: [],
+  fileFirst: false,
+  globOptions: {
+    cwd: './',
+    dot: true,
+    nonull: true,
+  },
+  includeCWD: true,
+  quote: "'",
+  targetExts: [
+    'ts'
+  ],
+  useSemicolon: false,
+  useTimestamp: false,
+  verbose: false,
+  withoutComment: true,
+  withoutBackupFile: true
+}
