+++ packages/main-interface/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@pokt-network/pocket-js-transactions",
+  "name": "@pokt-network/pocket-js-main-interface",
   "version": "0.0.1-rc",
-  "description": "PocketJS Transactions",
+  "description": "PocketJS Main interface",
   "main": "src/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
@@ -19,7 +19,7 @@
     "sdk",
     "typescript",
     "javascript",
-    "transactions"
+    "interface"
   ],
   "author": "Pocket Network",
   "license": "MIT",
@@ -30,9 +30,15 @@
     "typescript": "^3.7.2"
   },
   "dependencies": {
+    "@pokt-network/aat-js": "0.1.1-rc",
+    "@pokt-network/pocket-js-configuration": "file:../configuration",
+    "@pokt-network/pocket-js-http-provider": "file:../http-provider",
     "@pokt-network/pocket-js-keybase": "file:../keybase",
-    "@pokt-network/pocket-js-main-interface": "file:../main-interface",
+    "@pokt-network/pocket-js-relay-models": "file:../relay-models",
+    "@pokt-network/pocket-js-rpc": "file:../rpc",
     "@pokt-network/pocket-js-rpc-models": "file:../rpc-models",
+    "@pokt-network/pocket-js-session-manager": "file:../session-manager",
+    "@pokt-network/pocket-js-transactions": "file:../transactions",
     "@pokt-network/pocket-js-utils": "file:../utils"
   }
 }
