+++ packages/main-interface/tsconfig.json
@@ -1,0 +1,23 @@
+{
+  "compilerOptions": {
+    /* Basic Options */
+    "target": "es5",
+    "module": "commonjs",
+    "sourceMap": true,
+    "outDir": "./dist",
+    "strict": true,
+    "alwaysStrict": true,
+    "lib": ["es5", "es6", "dom"],
+    "downlevelIteration": true,
+    "declaration": true,
+    "esModuleInterop": true,
+    "watch": false
+  },
+  "exclude": [
+    "node_modules",
+    "**/*.spec.ts",
+    "lib",
+    "tests",
+    "build"
+  ]
+}
