+++ packages/main-interface/tslint.json
@@ -1,0 +1,29 @@
+{
+    "extends": ["tslint:recommended", "tslint-config-prettier"],
+    "rules": {
+        "ordered-imports": [false],
+        "no-bitwise": false,
+        "prefer-for-of": false,
+        "no-console": false,
+        "no-namespace": false,
+        "semicolon": false,
+        "no-unused-expression": false,
+        "only-arrow-functions": false,
+        "max-classes-per-file": [true, 2],
+        "forin": false,
+        "no-shadowed-variable": false,
+        "object-literal-shorthand": [false, "never"],
+        "object-literal-sort-keys": false,
+        "array-type": false,
+        "ban-types": false,
+        "no-var-keyword": false,
+        "prefer-const": false
+    },
+    "linterOptions": {
+        "exclude": [
+            "**/index.ts",
+            "lib/*"
+        ]
+    }
+ }
+
